+++ src/ReinforcePhase.cpp
@@ -14,7 +14,7 @@
 #include <string>
 
 int Reinforce::unitsForCountries(Player& player) {
-	int units = player.getCountry().size() / 3;
+	int units = (int)player.getCountry().size() / 3;
 	if (units < 3) {
 		units = 3;
 	}
